services/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dspy-ai>=3.0.3",
    "fastapi[standard]>=0.116.1",
    "openai>=1.107.2",
    "pinecone[grpc]>=7.3.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.13.0",
]
